Rename contans to constants
